backend/requirements.txt
# Core
Flask==3.0.0
Flask-SocketIO==5.3.4
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.0
psycopg2-binary==2.9.6
alembic==1.11.1

# WebSocket
eventlet==0.33.3
python-socketio==5.9.0

# Utilities
numpy==1.24.0
pandas==2.0.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.0
pytest-cov==4.1.0